Allow copied yaml to be edited in the create policy wizard (#2425)
* Always allow name/ns edits when creating a policy

Previously, if the user pasted in existing yaml with a uid, they could
not change the name or namespace via the form. Note that this commit
only fixes the form, not the editor.

Refs:
 - https://issues.redhat.com/browse/ACM-2188

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Protect fewer fields in create policy wizard

Previously, when `metadata.uid` was present in an object in the editor,
that field and its siblings would always be readonly. This prevented
changes to fields that should not be changed when editing a resource,
but also made it impossible to paste an existing resource into a create
wizard and edit those fields.

Now, this behavior is configurable in the SyncEditorProps. By default,
the old behavior will be used. The Create Policy wizard opts in and uses
the new behavior.

Refs:
 - https://issues.redhat.com/browse/ACM-2188

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Add test for edits on the create page

It should be possible to paste a "complete" policy into the editor now,
and still edit the name/namespace fields.

Refs:
 - https://issues.redhat.com/browse/ACM-2188

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Add test for conditionally protected ranges

When a YAML has a uid, whether or not it is protected from edits should
depend on whether the `editableUidSiblings` field is true or not.

Refs:
 - https://issues.redhat.com/browse/ACM-2188

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>