MGMT-13551 Move reconcilePause control to automationControlData (#2532)
* MGMT-13551 Move reconcilePause control to automationControlData

`reconcilePause` control flag is currently only used by `onChangeAutomationTemplate`
function inside automationControlData. The function relies on the control to be defined
in the controlData otherwise an error occurs when trying to set reconcilePause.active
value. Since the reconcilePause control was defined in clusterDetailsControlData, for the
wizards (e.g. CIM or AI) which don't use this step the error occured.

To fix this, we're moving the reconcilePause control to automationControlData which is only
controlData which actually uses that flag.

Signed-off-by: Jiri Tomasek <jtomasek@redhat.com>

* Update snapshots

Signed-off-by: Jiri Tomasek <jtomasek@redhat.com>

---------

Signed-off-by: Jiri Tomasek <jtomasek@redhat.com>