ACM-20098 Policy resources across clusters (#4525)
* Fix typo in DiscoveredPolicyTableItem

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Add navigation "tabs" to discovered policies page

Refs:
 - https://issues.redhat.com/browse/ACM-20098

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Move more work to the `useFetchPolicies` worker

Creating the `kyvernoPolicyReports` is not particularly intense, but
there may be some future calculations for related resources that could
benefit from being in the worker, if there are many policies. This
commit restructures the worker's response to fit that better.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Add a helper function for empty related resources

This will allow a new table to stay in sync with the existing one in
this situation.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Initial discovered policy resources page

The page is largely based on `PolicyTemplateDetails.tsx`. The work of
sorting through the search response is delegated to the worker thread.

Because not all the information we want to display on the page is
suitable to be saved in Search (for example, we want the 'Reason'
messages and `Diff` information when it is present on a
ConfigurationPolicy), the page uses ManagedClusterViews for each
displayed policy to get some more information. Those are only fired when
a related resource for that policy should be displayed, and they are
saved to prevent duplicated requests.

Refs:
 - https://issues.redhat.com/browse/ACM-20098

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Handle kyverno resources in discovered table

Refs:
 - https://issues.redhat.com/browse/ACM-20098

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Add related resources for certificate policy

Refs:
 - https://issues.redhat.com/browse/ACM-20098

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Add message for discovered gatekeeper resources

The search-collector does not track related resources for gatekeeper
constraints, and they will not be added this release, so this table
needed a special empty state. Those resources might be added in the
future.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Add DiscoveredResources tests and small updates

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Remove VAP from VAPB discovered resources table

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Remove discovered related resources tab for GK

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Use updater function for reasonCache

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

---------

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>